+++ src/main/java/com/example/spot/service/post/PostQueryService.java
@@ -14,4 +14,6 @@ public interface PostQueryService {
 
     PostAnnouncementResponse getPostAnnouncements();
 
+    CommentResponse getCommentsByPostId(Long postId);
+
 }
